api/requirements.txt
flask==3.1.0
flask-cors==5.0.1
gspread==6.1.4
google-auth==2.38.0
requests==2.32.3
python-dotenv==1.1.0
Flask-Limiter==3.8.0
waitress==3.0.0
